requirements.txt
numpy==1.19.2
matplotlib==3.3.2
cython==0.29.21
pyyaml==5.3.1
mypy==0.812
botorch==0.3.1
glfw==2.1.0
gpytorch==1.2.0
gym==0.9.3
mujoco-py==0.5.7
torch==1.6.0
torchvision==0.7.0

